develop Made Rental Agreement read endpoint populate the transient attributes.

+++ rentool/conf/create_db-0.0.1-SNAPSHOT.sql
@@ -45,7 +45,7 @@ CREATE TABLE IF NOT EXISTS rental_agrmnt
     rental_days         DECIMAL                        NOT NULL,
     checkout_date       DATE                           NOT NULL,
     daily_rental_charge FLOAT                          NOT NULL,
-    discount_percent    FLOAT                          NOT NULL,
+    discount_percent    INTEGER                        NOT NULL,
     state               VARCHAR(32) DEFAULT 'Proposed' NOT NULL,
     data_item           TEXT,
     FOREIGN KEY (tool_id) REFERENCES tool (id),
